feat: se añade crud para insumo y otros metodos de busqueda, venta y …

+++ src/main/java/com/burgerreal/burger_gestion/infrastructure/adapter/in/web/dto/venta/EditarVentaRequest.java
@@ -1,7 +1,9 @@
 package com.burgerreal.burger_gestion.infrastructure.adapter.in.web.dto.venta;
 
+import java.math.BigDecimal;
+
 public record EditarVentaRequest(
-        Long montoBruto,
-        Long costoInsumos,
+        BigDecimal montoBruto,
+        BigDecimal costoInsumos,
         Long metodoPagoId
 ) {}
